[fix] create payment gateway and payment gateway account from on_update of paypal settigns
